+++ portal-impl/src/com/liferay/portal/service/impl/LayoutSetPrototypeLocalServiceImpl.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -14,19 +14,26 @@
 
 package com.liferay.portal.service.impl;
 
+import com.liferay.portal.RequiredLayoutSetPrototypeException;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
+import com.liferay.portal.kernel.systemevent.SystemEvent;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.model.Group;
+import com.liferay.portal.model.GroupConstants;
 import com.liferay.portal.model.LayoutConstants;
 import com.liferay.portal.model.LayoutSetPrototype;
 import com.liferay.portal.model.ResourceConstants;
+import com.liferay.portal.model.SystemEventConstants;
+import com.liferay.portal.model.User;
 import com.liferay.portal.security.permission.PermissionCacheUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.base.LayoutSetPrototypeLocalServiceBaseImpl;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -38,13 +45,17 @@ import java.util.Map;
 public class LayoutSetPrototypeLocalServiceImpl
 	extends LayoutSetPrototypeLocalServiceBaseImpl {
 
+	@Override
 	public LayoutSetPrototype addLayoutSetPrototype(
 			long userId, long companyId, Map<Locale, String> nameMap,
-			String description, boolean active, boolean allowModifications,
-			boolean allowLayoutAddition, ServiceContext serviceContext)
+			String description, boolean active, boolean layoutsUpdateable,
+			ServiceContext serviceContext)
 		throws PortalException, SystemException {
 
 		// Layout set prototype
+
+		User user = userPersistence.findByPrimaryKey(userId);
+		Date now = new Date();
 
 		long layoutSetPrototypeId = counterLocalService.increment();
 
@@ -53,6 +64,10 @@ public class LayoutSetPrototypeLocalServiceImpl
 
 		layoutSetPrototype.setUuid(serviceContext.getUuid());
 		layoutSetPrototype.setCompanyId(companyId);
+		layoutSetPrototype.setUserId(userId);
+		layoutSetPrototype.setUserName(user.getFullName());
+		layoutSetPrototype.setCreateDate(serviceContext.getCreateDate(now));
+		layoutSetPrototype.setModifiedDate(serviceContext.getModifiedDate(now));
 		layoutSetPrototype.setNameMap(nameMap);
 		layoutSetPrototype.setDescription(description);
 		layoutSetPrototype.setActive(active);
@@ -61,13 +76,11 @@ public class LayoutSetPrototypeLocalServiceImpl
 			layoutSetPrototype.getSettingsProperties();
 
 		settingsProperties.put(
-			"allowModifications", String.valueOf(allowModifications));
-		settingsProperties.put(
-			"allowLayoutAdditions", String.valueOf(allowLayoutAddition));
+			"layoutsUpdateable", String.valueOf(layoutsUpdateable));
 
 		layoutSetPrototype.setSettingsProperties(settingsProperties);
 
-		layoutSetPrototypePersistence.update(layoutSetPrototype, false);
+		layoutSetPrototypePersistence.update(layoutSetPrototype);
 
 		// Resources
 
@@ -84,25 +97,41 @@ public class LayoutSetPrototypeLocalServiceImpl
 			"/template-" + layoutSetPrototype.getLayoutSetPrototypeId();
 
 		Group group = groupLocalService.addGroup(
-			userId, LayoutSetPrototype.class.getName(),
+			userId, GroupConstants.DEFAULT_PARENT_GROUP_ID,
+			LayoutSetPrototype.class.getName(),
 			layoutSetPrototype.getLayoutSetPrototypeId(),
-			layoutSetPrototype.getName(LocaleUtil.getDefault()), null, 0,
-			friendlyURL, false, true, serviceContext);
+			GroupConstants.DEFAULT_LIVE_GROUP_ID,
+			layoutSetPrototype.getName(LocaleUtil.getDefault()), null, 0, true,
+			GroupConstants.DEFAULT_MEMBERSHIP_RESTRICTION, friendlyURL, false,
+			true, serviceContext);
 
-		layoutLocalService.addLayout(
-			userId, group.getGroupId(), true,
-			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, "home", null, null,
-			LayoutConstants.TYPE_PORTLET, false, "/home", false,
-			serviceContext);
+		if (GetterUtil.getBoolean(
+				serviceContext.getAttribute("addDefaultLayout"), true)) {
+
+			layoutLocalService.addLayout(
+				userId, group.getGroupId(), true,
+				LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, "home", null, null,
+				LayoutConstants.TYPE_PORTLET, false, "/home", serviceContext);
+		}
 
 		return layoutSetPrototype;
 	}
 
 	@Override
-	public void deleteLayoutSetPrototype(LayoutSetPrototype layoutSetPrototype)
+	@SystemEvent(
+		action = SystemEventConstants.ACTION_SKIP,
+		type = SystemEventConstants.TYPE_DELETE)
+	public LayoutSetPrototype deleteLayoutSetPrototype(
+			LayoutSetPrototype layoutSetPrototype)
 		throws PortalException, SystemException {
 
 		// Group
+
+		if (layoutSetPersistence.countByLayoutSetPrototypeUuid(
+				layoutSetPrototype.getUuid()) > 0) {
+
+			throw new RequiredLayoutSetPrototypeException();
+		}
 
 		Group group = layoutSetPrototype.getGroup();
 
@@ -123,25 +152,66 @@ public class LayoutSetPrototypeLocalServiceImpl
 		// Permission cache
 
 		PermissionCacheUtil.clearCache();
+
+		return layoutSetPrototype;
 	}
 
 	@Override
-	public void deleteLayoutSetPrototype(long layoutSetPrototypeId)
+	public LayoutSetPrototype deleteLayoutSetPrototype(
+			long layoutSetPrototypeId)
 		throws PortalException, SystemException {
 
 		LayoutSetPrototype layoutSetPrototype =
 			layoutSetPrototypePersistence.findByPrimaryKey(
 				layoutSetPrototypeId);
 
-		deleteLayoutSetPrototype(layoutSetPrototype);
+		return deleteLayoutSetPrototype(layoutSetPrototype);
 	}
 
+	@Override
+	public void deleteNondefaultLayoutSetPrototypes(long companyId)
+		throws PortalException, SystemException {
+
+		long defaultUserId = userLocalService.getDefaultUserId(companyId);
+
+		List<LayoutSetPrototype> layoutSetPrototypes =
+			layoutSetPrototypePersistence.findByCompanyId(companyId);
+
+		for (LayoutSetPrototype layoutSetPrototype : layoutSetPrototypes) {
+			if (layoutSetPrototype.getUserId() != defaultUserId) {
+				deleteLayoutSetPrototype(layoutSetPrototype);
+			}
+		}
+	}
+
+	/**
+	 * @deprecated As of 6.2.0, replaced by {@link
+	 *             #getLayoutSetPrototypeByUuidAndCompanyId(String, long)}
+	 */
+	@Override
 	public LayoutSetPrototype getLayoutSetPrototypeByUuid(String uuid)
 		throws PortalException, SystemException {
 
 		return layoutSetPrototypePersistence.findByUuid_First(uuid, null);
 	}
 
+	@Override
+	public LayoutSetPrototype getLayoutSetPrototypeByUuidAndCompanyId(
+			String uuid, long companyId)
+		throws PortalException, SystemException {
+
+		return layoutSetPrototypePersistence.findByUuid_C_First(
+			uuid, companyId, null);
+	}
+
+	@Override
+	public List<LayoutSetPrototype> getLayoutSetPrototypes(long companyId)
+		throws SystemException {
+
+		return layoutSetPrototypePersistence.findByCompanyId(companyId);
+	}
+
+	@Override
 	public List<LayoutSetPrototype> search(
 			long companyId, Boolean active, int start, int end,
 			OrderByComparator obc)
@@ -157,6 +227,7 @@ public class LayoutSetPrototypeLocalServiceImpl
 		}
 	}
 
+	@Override
 	public int searchCount(long companyId, Boolean active)
 		throws SystemException {
 
@@ -168,10 +239,11 @@ public class LayoutSetPrototypeLocalServiceImpl
 		}
 	}
 
+	@Override
 	public LayoutSetPrototype updateLayoutSetPrototype(
 			long layoutSetPrototypeId, Map<Locale, String> nameMap,
-			String description, boolean active, boolean allowModifications,
-			boolean allowLayoutAddition, ServiceContext serviceContext)
+			String description, boolean active, boolean layoutsUpdateable,
+			ServiceContext serviceContext)
 		throws PortalException, SystemException {
 
 		// Layout set prototype
@@ -180,6 +252,8 @@ public class LayoutSetPrototypeLocalServiceImpl
 			layoutSetPrototypePersistence.findByPrimaryKey(
 				layoutSetPrototypeId);
 
+		layoutSetPrototype.setModifiedDate(
+			serviceContext.getModifiedDate(new Date()));
 		layoutSetPrototype.setNameMap(nameMap);
 		layoutSetPrototype.setDescription(description);
 		layoutSetPrototype.setActive(active);
@@ -188,13 +262,11 @@ public class LayoutSetPrototypeLocalServiceImpl
 			layoutSetPrototype.getSettingsProperties();
 
 		settingsProperties.put(
-			"allowModifications", String.valueOf(allowModifications));
-		settingsProperties.put(
-			"allowLayoutAdditions", String.valueOf(allowLayoutAddition));
+			"layoutsUpdateable", String.valueOf(layoutsUpdateable));
 
 		layoutSetPrototype.setSettingsProperties(settingsProperties);
 
-		layoutSetPrototypePersistence.update(layoutSetPrototype, false);
+		layoutSetPrototypePersistence.update(layoutSetPrototype);
 
 		// Group
 
@@ -203,11 +275,12 @@ public class LayoutSetPrototypeLocalServiceImpl
 
 		group.setName(layoutSetPrototype.getName(LocaleUtil.getDefault()));
 
-		groupPersistence.update(group, false);
+		groupPersistence.update(group);
 
 		return layoutSetPrototype;
 	}
 
+	@Override
 	public LayoutSetPrototype updateLayoutSetPrototype(
 			long layoutSetPrototypeId, String settings)
 		throws PortalException, SystemException {
@@ -218,9 +291,10 @@ public class LayoutSetPrototypeLocalServiceImpl
 			layoutSetPrototypePersistence.findByPrimaryKey(
 				layoutSetPrototypeId);
 
+		layoutSetPrototype.setModifiedDate(new Date());
 		layoutSetPrototype.setSettings(settings);
 
-		layoutSetPrototypePersistence.update(layoutSetPrototype, false);
+		layoutSetPrototypePersistence.update(layoutSetPrototype);
 
 		// Group
 
@@ -243,7 +317,7 @@ public class LayoutSetPrototypeLocalServiceImpl
 
 		group.setTypeSettings(typeSettingsProperties.toString());
 
-		groupPersistence.update(group, false);
+		groupPersistence.update(group);
 
 		return layoutSetPrototype;
 	}
