+++ portal-impl/src/com/liferay/portlet/wiki/action/EditNodeAction.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -14,21 +14,33 @@
 
 package com.liferay.portlet.wiki.action;
 
+import com.liferay.portal.kernel.portlet.LiferayPortletConfig;
 import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.servlet.SessionMessages;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextFactory;
 import com.liferay.portal.struts.PortletAction;
+import com.liferay.portal.theme.ThemeDisplay;
+import com.liferay.portal.util.WebKeys;
+import com.liferay.portlet.trash.util.TrashUtil;
 import com.liferay.portlet.wiki.DuplicateNodeNameException;
 import com.liferay.portlet.wiki.NoSuchNodeException;
 import com.liferay.portlet.wiki.NodeNameException;
+import com.liferay.portlet.wiki.RequiredNodeException;
 import com.liferay.portlet.wiki.model.WikiNode;
+import com.liferay.portlet.wiki.service.WikiNodeLocalServiceUtil;
 import com.liferay.portlet.wiki.service.WikiNodeServiceUtil;
 import com.liferay.portlet.wiki.util.WikiCacheThreadLocal;
 import com.liferay.portlet.wiki.util.WikiCacheUtil;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -48,8 +60,9 @@ public class EditNodeAction extends PortletAction {
 
 	@Override
 	public void processAction(
-			ActionMapping mapping, ActionForm form, PortletConfig portletConfig,
-			ActionRequest actionRequest, ActionResponse actionResponse)
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, ActionRequest actionRequest,
+			ActionResponse actionResponse)
 		throws Exception {
 
 		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
@@ -59,7 +72,15 @@ public class EditNodeAction extends PortletAction {
 				updateNode(actionRequest);
 			}
 			else if (cmd.equals(Constants.DELETE)) {
-				deleteNode(actionRequest);
+				deleteNode(
+					(LiferayPortletConfig)portletConfig, actionRequest, false);
+			}
+			else if (cmd.equals(Constants.MOVE_TO_TRASH)) {
+				deleteNode(
+					(LiferayPortletConfig)portletConfig, actionRequest, true);
+			}
+			else if (cmd.equals(Constants.RESTORE)) {
+				restoreNode(actionRequest);
 			}
 			else if (cmd.equals(Constants.SUBSCRIBE)) {
 				subscribeNode(actionRequest);
@@ -74,14 +95,14 @@ public class EditNodeAction extends PortletAction {
 			if (e instanceof NoSuchNodeException ||
 				e instanceof PrincipalException) {
 
-				SessionErrors.add(actionRequest, e.getClass().getName());
+				SessionErrors.add(actionRequest, e.getClass());
 
 				setForward(actionRequest, "portlet.wiki.error");
 			}
 			else if (e instanceof DuplicateNodeNameException ||
 					 e instanceof NodeNameException) {
 
-				SessionErrors.add(actionRequest, e.getClass().getName());
+				SessionErrors.add(actionRequest, e.getClass());
 			}
 			else {
 				throw e;
@@ -91,8 +112,9 @@ public class EditNodeAction extends PortletAction {
 
 	@Override
 	public ActionForward render(
-			ActionMapping mapping, ActionForm form, PortletConfig portletConfig,
-			RenderRequest renderRequest, RenderResponse renderResponse)
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, RenderRequest renderRequest,
+			RenderResponse renderResponse)
 		throws Exception {
 
 		try {
@@ -106,33 +128,82 @@ public class EditNodeAction extends PortletAction {
 			if (e instanceof NoSuchNodeException ||
 				e instanceof PrincipalException) {
 
-				SessionErrors.add(renderRequest, e.getClass().getName());
+				SessionErrors.add(renderRequest, e.getClass());
 
-				return mapping.findForward("portlet.wiki.error");
+				return actionMapping.findForward("portlet.wiki.error");
 			}
 			else {
 				throw e;
 			}
 		}
 
-		return mapping.findForward(
+		return actionMapping.findForward(
 			getForward(renderRequest, "portlet.wiki.edit_node"));
 	}
 
-	protected void deleteNode(ActionRequest actionRequest) throws Exception {
+	protected void deleteNode(
+			LiferayPortletConfig liferayPortletConfig,
+			ActionRequest actionRequest, boolean moveToTrash)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		int nodeCount = WikiNodeLocalServiceUtil.getNodesCount(
+			themeDisplay.getScopeGroupId());
+
+		if (nodeCount == 1) {
+			SessionErrors.add(actionRequest, RequiredNodeException.class);
+
+			return;
+		}
+
 		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
 
 		String oldName = getNodeName(nodeId);
 
 		WikiCacheThreadLocal.setClearCache(false);
 
-		WikiNodeServiceUtil.deleteNode(nodeId);
+		String deleteEntryTitle = null;
+
+		if (moveToTrash) {
+			WikiNode node = WikiNodeServiceUtil.moveNodeToTrash(nodeId);
+
+			deleteEntryTitle = node.getName();
+		}
+		else {
+			WikiNodeServiceUtil.deleteNode(nodeId);
+		}
 
 		WikiCacheUtil.clearCache(nodeId);
 
 		WikiCacheThreadLocal.setClearCache(true);
 
 		updatePreferences(actionRequest, oldName, StringPool.BLANK);
+
+		if (moveToTrash) {
+			Map<String, String[]> data = new HashMap<String, String[]>();
+
+			data.put(
+				"deleteEntryClassName",
+				new String[] {WikiNode.class.getName()});
+
+			if (Validator.isNotNull(deleteEntryTitle)) {
+				data.put(
+					"deleteEntryTitle",
+					new String[] {
+						TrashUtil.getOriginalTitle(deleteEntryTitle)});
+			}
+
+			data.put("restoreEntryIds", new String[] {String.valueOf(nodeId)});
+
+			SessionMessages.add(
+				actionRequest,
+				liferayPortletConfig.getPortletId() +
+					SessionMessages.KEY_SUFFIX_DELETE_SUCCESS_DATA, data);
+
+			hideDefaultSuccessMessage(liferayPortletConfig, actionRequest);
+		}
 	}
 
 	protected String getNodeName(long nodeId) throws Exception {
@@ -141,9 +212,16 @@ public class EditNodeAction extends PortletAction {
 		return node.getName();
 	}
 
-	protected void subscribeNode(ActionRequest actionRequest)
-		throws Exception {
+	protected void restoreNode(ActionRequest actionRequest) throws Exception {
+		long[] restoreEntryIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "restoreEntryIds"), 0L);
 
+		for (long restoreEntryId : restoreEntryIds) {
+			WikiNodeServiceUtil.restoreNodeFromTrash(restoreEntryId);
+		}
+	}
+
+	protected void subscribeNode(ActionRequest actionRequest) throws Exception {
 		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
 
 		WikiNodeServiceUtil.subscribeNode(nodeId);
@@ -189,22 +267,22 @@ public class EditNodeAction extends PortletAction {
 			ActionRequest actionRequest, String oldName, String newName)
 		throws Exception {
 
-		PortletPreferences preferences = actionRequest.getPreferences();
+		PortletPreferences portletPreferences = actionRequest.getPreferences();
 
-		String hiddenNodes = preferences.getValue(
+		String hiddenNodes = portletPreferences.getValue(
 			"hiddenNodes", StringPool.BLANK);
-		String visibleNodes = preferences.getValue(
+		String visibleNodes = portletPreferences.getValue(
 			"visibleNodes", StringPool.BLANK);
 
 		String regex = oldName + ",?";
 
-		preferences.setValue(
+		portletPreferences.setValue(
 			"hiddenNodes", hiddenNodes.replaceFirst(regex, newName));
-		preferences.setValue(
+		portletPreferences.setValue(
 			"visibleNodes",
 			visibleNodes.replaceFirst(regex, newName));
 
-		preferences.store();
+		portletPreferences.store();
 	}
 
 }
